# feat: sanitize structured metadata during ingestion in the distributor