--- a/pkg/distributor/distributor.go
+++ b/pkg/distributor/distributor.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"flag"
 	"fmt"
+	otlptranslate "github.com/prometheus/prometheus/storage/remote/otlptranslator/prometheus"
 	"math"
 	"net/http"
 	"slices"
@@ -12,6 +13,7 @@ import (
 	"strings"
 	"sync"
 	"time"
+	"unicode/utf8"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
@@ -66,6 +68,14 @@ const (
 var (
 	maxLabelCacheSize = 100000
 	rfStats           = analytics.NewInt("distributor_replication_factor")
+
+	// the rune error replacement is rejected by Prometheus hence replacing them with space.
+	removeInvalidUtf = func(r rune) rune {
+		if r == utf8.RuneError {
+			return 32 // rune value for space
+		}
+		return r
+	}
 )
 
 // Config for a Distributor.
@@ -519,6 +529,12 @@ func (d *Distributor) Push(ctx context.Context, req *logproto.PushRequest) (*log
 				structuredMetadata := logproto.FromLabelAdaptersToLabels(entry.StructuredMetadata)
 				if shouldDiscoverLevels {
 					logLevel, ok := levelDetector.extractLogLevel(lbs, structuredMetadata, entry)
+					for i := range entry.StructuredMetadata {
+						structuredMetadata[i].Name = otlptranslate.NormalizeLabel(structuredMetadata[i].Name)
+						if strings.ContainsRune(structuredMetadata[i].Value, utf8.RuneError) {
+							structuredMetadata[i].Value = strings.Map(removeInvalidUtf, structuredMetadata[i].Value)
+						}
+					}
 					if ok {
 						entry.StructuredMetadata = append(entry.StructuredMetadata, logLevel)
 					}