--- a/pkg/distributor/distributor.go
+++ b/pkg/distributor/distributor.go
@@ -528,12 +528,10 @@ func (d *Distributor) Push(ctx context.Context, req *logproto.PushRequest) (*log
 				}
 
 				structuredMetadata := logproto.FromLabelAdaptersToLabels(entry.StructuredMetadata)
-				if len(entry.StructuredMetadata) > 0 {
-					for i := range entry.StructuredMetadata {
-						structuredMetadata[i].Name = otlptranslate.NormalizeLabel(structuredMetadata[i].Name)
-						if strings.ContainsRune(structuredMetadata[i].Value, utf8.RuneError) {
-							structuredMetadata[i].Value = strings.Map(removeInvalidUtf, structuredMetadata[i].Value)
-						}
+				for i := range entry.StructuredMetadata {
+					structuredMetadata[i].Name = otlptranslate.NormalizeLabel(structuredMetadata[i].Name)
+					if strings.ContainsRune(structuredMetadata[i].Value, utf8.RuneError) {
+						structuredMetadata[i].Value = strings.Map(removeInvalidUtf, structuredMetadata[i].Value)
 					}
 				}
 				if shouldDiscoverLevels {