@cstyan 2024-11-29
> LGTM
> 
> Did you find why we have multiple detected_level ?

I did, it was just related to reuse of the same `logproto.PushRequest` for each iteration of the benchmark in `b.N`, since the distributor `Push` function actually modifies the structured metadata in the struct to add the `detected_level`.

I modified the benchmark to create a new write request for each iteration of `b.N`. Unfortunately it seems like the `b.ResetTimer` or start/stop timer calls break the benchmark in some way that I wasn't able to track down, related to the distributor/mock ingester health check in some way since that's what the panic stack trace shows.